# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 Swagger UI",
        "description": "Интерактивная документация HTTP API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Текущий хост приложения"
        }
    ],
    "paths": {
...
        "/api/orgs": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Список организаций",
                "description": "Постраничный список. По умолчанию сортировка по `id` по убыванию (сначала записи с большим id). Поиск `search` — подстрока в названии (без учёта регистра). Параметры `filter[...]` задают точное совпадение по колонке.",
                "operationId": "0ca58ec28cf5b4404936c1a0416e4ba7",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Размер страницы (1–1000), по умолчанию 15",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15,
                            "maximum": 1000,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Колонка сортировки",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "id",
                            "enum": [
                                "id",
                                "name",
                                "slug",
                                "about",
                                "logo",
                                "website",
                                "email",
                                "phone",
                                "address",
                                "city",
                                "status",
                                "created_at",
                                "updated_at"
                            ]
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Направление сортировки (по умолчанию desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Подстрока поиска в названии организации (без учёта регистра)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Acme"
                        }
                    },
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[slug]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[about]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[logo]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[website]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[email]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "email"
                        }
                    },
                    {
                        "name": "filter[phone]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[address]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[city]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[status]",
                        "in": "query",
                        "description": "Точное значение статуса",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "new",
                                "enabled",
                                "deleted"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Страница списка организаций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "data",
                                        "links",
                                        "meta"
                                    ],
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Org"
                                            }
                                        },
                                        "links": {
                                            "description": "Ссылки пагинации (first, last, prev, next)",
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "nullable": true
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "path": {
                                                    "type": "string"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "nullable": true
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            },
            "post": {
                "tags": [
                    "Организации"
                ],
                "summary": "Создать организацию",
                "description": "Создаёт новую организацию со статусом `new`. Поле slug необязательно: при отсутствии генерируется из name.",
                "operationId": "646eb64ec7dcae8ebb86ac7d278835ea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrgStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Организация создана",
                        "headers": {
                            "Location": {
                                "description": "URL созданной организации (GET)",
                                "schema": {
                                    "type": "string",
                                    "example": "/api/orgs/1"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "data"
                                    ],
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Org"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/orgs/{org}": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Получить организацию по идентификатору",
                "description": "Возвращает одну организацию по числовому первичному ключу.",
                "operationId": "258ea0f0b99ef4680003754631f2c79d",
                "parameters": [
                    {
                        "name": "org",
                        "in": "path",
                        "description": "Числовой идентификатор организации (первичный ключ)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Организация найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "data"
                                    ],
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Org"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Организация с указанным id не найдена"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Организации"
                ],
                "summary": "Обновить организацию (частично)",
                "description": "Частичное обновление организации. Поля являются опциональными: если поле не передано — оно не изменяется. Если поле передано как null — значение очищается (становится null) для nullable-колонок. Для slug: null означает регенерацию slug из name.",
                "operationId": "ad08e5d3936181129c01a88920a6cbda",
                "parameters": [
                    {
                        "name": "org",
                        "in": "path",
                        "description": "Числовой идентификатор организации (первичный ключ)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "minimum": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrgUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Организация обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "data"
                                    ],
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Org"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Организация с указанным id не найдена"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Org": {
                "required": [
                    "id",
                    "name",
                    "slug",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Acme Inc"
                    },
                    "slug": {
                        "type": "string",
                        "example": "acme-inc"
                    },
                    "about": {
                        "type": "string",
                        "nullable": true
                    },
                    "logo": {
                        "type": "string",
                        "nullable": true
                    },
                    "website": {
                        "type": "string",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "nullable": true
                    },
                    "city": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "description": "Статус организации",
                        "type": "string",
                        "nullable": true,
                        "enum": [
                            "new",
                            "enabled",
                            "deleted"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "OrgStoreRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Acme Inc",
                        "maxLength": 255
                    },
                    "slug": {
                        "description": "Уникальный slug; если не указан, будет сгенерирован из name",
                        "type": "string",
                        "example": "acme-inc",
                        "nullable": true,
                        "maxLength": 255
                    },
                    "about": {
                        "type": "string",
                        "nullable": true
                    },
                    "logo": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 255
                    },
                    "website": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true,
                        "maxLength": 255
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "nullable": true,
                        "maxLength": 255
                    },
                    "phone": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 255
                    },
                    "address": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 255
                    },
                    "city": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 255
                    }
                },
                "type": "object"
            },
            "OrgUpdateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Acme Inc",
                        "maxLength": 255
                    },
                    "slug": {
                        "description": "Уникальный slug. Если передать null — slug будет регенерирован из name.",
                        "type": "string",
                        "example": "acme-inc",
                        "nullable": true,
                        "maxLength": 255
                    },
                    "about": {
                        "type": "string",
                        "nullable": true
                    },
                    "logo": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 255
                    },
                    "website": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true,
                        "maxLength": 255
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "nullable": true,
                        "maxLength": 255
                    },
                    "phone": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 255
                    },
                    "address": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 255
                    },
                    "city": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 255
                    },
                    "status": {
                        "description": "Статус организации",
                        "type": "string",
                        "nullable": true,
                        "enum": [
                            "new",
                            "enabled",
                            "deleted"
                        ]
                    }
                },
                "type": "object"
            }
        }
    },
...
}